weather-exporter/requirements.txt
prometheus_client==0.20.0
requests==2.31.0
kafka-python==2.0.2